Caesar Cipher in Cryptography
The Caesar Cipher technique is one of the earliest and simplest method
of encryption technique. It’s simply a type of substitution cipher,
i.e., each letter of a given text is replaced by a letter some fixed
number of positions down the alphabet. For example with a shift of 1, A
would be replaced by B, B would become C, and so on. The method is
apparently named after Julius Caesar, who apparently used it to
communicate with his officials.